feat : 결제 연동 구현

+++ src/main/java/nbc_final/gathering/domain/ad/dto/response/AdCreateResponseDto.java
@@ -9,11 +9,11 @@ public class AdCreateResponseDto {
     private final Long adId;
     private final LocalDate startDate;
     private final LocalDate endDate;
-    private final int totalAmount; // 총 금액
+    private final long totalAmount; // 총 금액
     private final String orderName; // 광고 형태 추가
     private final long adDuration; // 광고 기간
 
-    public AdCreateResponseDto(Long adId, LocalDate startDate, LocalDate endDate, int totalAmount, String orderName, long adDuration) {
+    public AdCreateResponseDto(Long adId, LocalDate startDate, LocalDate endDate, long totalAmount, String orderName, long adDuration) {
         this.adId = adId;
         this.startDate = startDate;
         this.endDate = endDate;
@@ -22,7 +22,9 @@ public class AdCreateResponseDto {
         this.adDuration = adDuration;
     }
 
-    public static AdCreateResponseDto of(Long adId, LocalDate startDate, LocalDate endDate, int totalAmount, String orderName, long adDuration) {
+    public static AdCreateResponseDto of(Long adId, LocalDate startDate, LocalDate endDate, long totalAmount, String orderName, long adDuration) {
         return new AdCreateResponseDto(adId, startDate, endDate, totalAmount, orderName, adDuration);
     }
+
+
 }
